📫 dispatch message helpers
