+++ packages/scms-server/src/backend/loaders/jobs/dispatch/base.ts
@@ -1,0 +1,6 @@
+import { uuidv7 } from 'uuidv7';
+
+/** Shared job id generator for dispatch factories. */
+export function newDispatchJobId(): string {
+  return uuidv7();
+}
